requirements.txt
streamlit>=1.40.0
tensorflow
openai-whisper
gTTS
pydub
epitran
jiwer
editdistance
numpy
pandas
